requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pdfkit==1.0.0
streamlit==1.40.1
python-dotenv==1.0.1
huggingface_hub==0.30.2
faiss-cpu==1.8.0.post1
langchain==0.2.17
PyPDF2==3.0.1
